markers are now saved in the database
The markers table has been added to the database.

Server/db.js was created so that the database connection can be exported.

Server/marker-repository.js was created where the queries to be performed related to the markers are located.
in Server/index.js added GET, POST and DELETE routes for the markers